backend/Models/segmentation/requirements.txt
# VirtualFit Segmentation Module Dependencies
# Install with: pip install -r requirements.txt

# Core Dependencies
numpy>=1.24.0
Pillow>=10.0.0
torch>=2.0.0
torchvision>=0.15.0

# SAM (Segment Anything Model)
segment-anything>=1.0

# Image Processing
opencv-python>=4.8.0

# For pose estimation (Phase 2)
mediapipe>=0.10.0
